+++ src/frontend/apps/e2e/__tests__/app-impress/doc-tree.spec.ts
@@ -9,6 +9,7 @@ import {
 } from './utils-common';
 import { addNewMember } from './utils-share';
 import {
+  addChild,
   clickOnAddRootSubPage,
   createRootSubPage,
   getTreeRow,
@@ -17,6 +18,137 @@ import {
 test.describe('Doc Tree', () => {
   test.beforeEach(async ({ page }) => {
     await page.goto('/');
+  });
+
+  test('check the tree pagination', async ({ page, browserName }) => {
+    await page.route(/.*\/documents\/.*\/children\//, async (route) => {
+      const request = route.request();
+      const url = new URL(request.url());
+      const pageId = url.searchParams.get('page') ?? '1';
+
+      const response = {
+        count: 40,
+        next: `http://localhost:8071/api/v1.0/documents/anything/children/?page=${parseInt(pageId) + 1}`,
+        previous:
+          parseInt(pageId) > 1
+            ? `http://localhost:8071/api/v1.0/documents/anything/children/?page=${parseInt(pageId) - 1}`
+            : null,
+        results: Array.from({ length: 20 }, (_, i) => ({
+          id: `doc-child-${pageId}-${i}`,
+          abilities: {
+            accesses_manage: true,
+            accesses_view: true,
+            ai_proxy: true,
+            ai_transform: true,
+            ai_translate: true,
+            attachment_upload: true,
+            media_check: true,
+            can_edit: true,
+            children_list: true,
+            children_create: true,
+            collaboration_auth: true,
+            comment: true,
+            content: true,
+            cors_proxy: true,
+            descendants: true,
+            destroy: true,
+            duplicate: true,
+            favorite: true,
+            link_configuration: true,
+            invite_owner: true,
+            mask: true,
+            move: true,
+            partial_update: true,
+            restore: true,
+            retrieve: true,
+            media_auth: true,
+            link_select_options: {
+              restricted: null,
+              authenticated: ['reader', 'commenter', 'editor'],
+              public: ['reader', 'commenter', 'editor'],
+            },
+            tree: true,
+            update: true,
+            versions_destroy: true,
+            versions_list: true,
+            versions_retrieve: true,
+            search: true,
+          },
+          ancestors_link_reach: 'restricted',
+          ancestors_link_role: null,
+          computed_link_reach: 'restricted',
+          computed_link_role: null,
+          created_at: '2026-03-27T14:44:12.398544Z',
+          creator: '40d339e9-cd97-4fdc-b65f-0a809c7e2db9',
+          deleted_at: null,
+          depth: 3,
+          excerpt: null,
+          is_favorite: false,
+          link_role: 'reader',
+          link_reach: 'restricted',
+          nb_accesses_ancestors: 1,
+          nb_accesses_direct: 0,
+          numchild: 0,
+          path: `000000p00000010000001-${pageId}-${i}`,
+          title: `doc-child-${pageId}-${i}`,
+          updated_at: '2026-03-27T14:44:26.691903Z',
+          user_role: 'owner',
+        })),
+      };
+
+      if (request.method().includes('GET')) {
+        await route.fulfill({
+          json: response,
+        });
+      } else {
+        await route.continue();
+      }
+    });
+
+    const [title] = await createDoc(
+      page,
+      'doc-tree-pagination',
+      browserName,
+      1,
+    );
+
+    const pageParentUrl = page.url();
+
+    const titleChild = await addChild({
+      page,
+      browserName,
+      docParent: title,
+      docName: 'doc-tree-pagination-child',
+    });
+
+    await addChild({
+      page,
+      browserName,
+      docParent: titleChild,
+      docName: 'doc-tree-pagination-child-2',
+    });
+
+    await page.goto(pageParentUrl);
+
+    await verifyDocName(page, title);
+
+    const docTree = page.getByTestId('doc-tree');
+    await expect(docTree).toBeVisible();
+    await docTree.getByText('keyboard_arrow_right').click();
+    await docTree
+      .getByRole('button', {
+        name: `Open document ${titleChild}`,
+      })
+      .click();
+
+    await expect(docTree.getByText('doc-child-1-19')).toBeVisible();
+    await expect(docTree.locator('.c__spinner')).toBeVisible();
+    await docTree.getByText('doc-child-1-19').hover();
+    await expect(
+      docTree.getByText('doc-child-2-1', {
+        exact: true,
+      }),
+    ).toBeVisible();
   });
 
   test('check the reorder of sub pages', async ({ page, browserName }) => {